+++ CMake/CPackConfig.cmake
@@ -1,6 +1,5 @@
-set(CPACK_PACKAGE_CONTACT "Pablo Toharia <pablo.toharia@urjc.es>")
-set(CPACK_PACKAGE_DESCRIPTION_FILE "${CMAKE_SOURCE_DIR}/README.md")
-set(CPACK_PACKAGE_LICENSE "Proprietary")
+set(CPACK_PACKAGE_CONTACT "Pablo Toharia <pablo.toharia@upm.es>")
+set(CPACK_PACKAGE_DESCRIPTION_FILE "${PROJECT_SOURCE_DIR}/README.md")
 
 set(CPACK_DEBIAN_PACKAGE_DEPENDS "")